requirements.txt
alembic==1.7.5
Flask==2.0.2
flask_migrate==3.1.0
flask_mysqldb==0.2.0
flask_sqlalchemy==2.5.1
PyJWT==2.3.0
python_bcrypt==0.3.2
SQLAlchemy==1.4.31
Werkzeug==1.0.0
